pyproject.toml
[project]
name = "dionysus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "chainlit",
    "fastapi>=0.115.0",
    "graphiti-core",
    "hephaestus",
    "langgraph",
    "langgraph-checkpoint-postgres>=3.0.4",
    "mem0ai>=1.0.2",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic",
    "pydantic-settings>=2.12.0",
    "python-socketio>=5.12.0",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "uvicorn[standard]>=0.34.0",
]

[tool.uv.sources]
hephaestus = { path = "../hephaestus", editable = true }
